requirements.txt
einops==0.8.1
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
pingouin==0.5.5
scikit_learn==1.7.2
scipy==1.16.2
torch==2.7.1
torcheval==0.0.7
torchvision==0.23.0
tqdm==4.67.1
vit_pytorch==1.12.4
